feat: Refactor tab and section management by simplifying API functions and enhancing content creation hooks
