+++ src/components/shared/class-guides/editor/components/tabs/editor-dialog.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useState, useEffect, useRef, useCallback } from 'react';
+import React from 'react';
 import {
   Dialog,
   DialogContent,
@@ -14,7 +14,7 @@ import { Label } from '@root/components/ui/label';
 import { Trash2 } from 'lucide-react';
 import { cn } from '@root/lib/utils';
 import { TabData } from '@root/@types/prisma';
-import toast from 'react-hot-toast';
+import { useEditorDialog } from '@root/components/hooks/guide/edit/tab-editor/use-tab-editor-dialog';
 
 interface EditorDialogProps {
   editingTab: TabData | null;
@@ -39,51 +39,14 @@ export const EditorDialog: React.FC<EditorDialogProps> = React.memo(
     onDelete,
     className,
   }) => {
-    const [isOpen, setIsOpen] = useState(!!editingTab);
-    const [isDeleting, setIsDeleting] = useState(false);
-    const labelRef = useRef<HTMLInputElement>(null);
-
-    useEffect(() => {
-      setIsOpen(!!editingTab);
-      if (editingTab && labelRef.current) {
-        labelRef.current.focus();
-      }
-    }, [editingTab]);
-
-    const validateForm = useCallback(() => {
-      if (!editForm.label.trim()) {
-        toast.error('Название таба не может быть пустым');
-        return false;
-      }
-      return true;
-    }, [editForm.label]);
-
-    const handleSave = () => {
-      if (!validateForm()) return;
-      onSave();
-      setIsOpen(false);
-      setTimeout(() => onClose(), 300);
-    };
-
-    const handleDelete = async () => {
-      setIsDeleting(true);
-      try {
-        await onDelete();
-        setIsDeleting(false);
-        setIsOpen(false);
-        setTimeout(() => onClose(), 300);
-      } catch {
-        setIsDeleting(false);
-        toast.error('Ошибка при удалении таба');
-      }
-    };
-
-    const handleOpenChange = (open: boolean) => {
-      if (!open) {
-        setIsOpen(false);
-        setTimeout(() => onClose(), 300);
-      }
-    };
+    const {
+      isOpen,
+      isDeleting,
+      labelRef,
+      handleSave,
+      handleDelete,
+      handleOpenChange,
+    } = useEditorDialog(editingTab, editForm, onClose, onSave, onDelete);
 
     return (
       <Dialog open={isOpen} onOpenChange={handleOpenChange}>
